backend/requirements.txt
# Core ML and Data Processing
tensorflow>=2.15.0
tensorflow-datasets>=4.9.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# API Framework (minimal for now)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Development and utilities
jupyter==1.1.1
tqdm==4.66.1
python-dotenv==1.0.0

# Data visualization
seaborn==0.12.2